requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
celery==5.4.0
langchain-community==0.3.1
langchain>=0.3.1
langchain_groq==0.2.0
langgraph==0.2.21
rabbitmq-client==2.4.0
requests==2.32.3
flower==2.0.1
pydantic-settings==2.5.2
fpdf2==2.7.9
duckduckgo-search==6.2.13